overall cleanup
