--- crates/benches/src/account_manager.rs
@@ -1,83 +1,0 @@
-use std::sync::Arc;
-
-use futures::future::join_all;
-use katana_core::accounts::DevAccountGenerator;
-use starknet::accounts::{Account, ConnectedAccount, ExecutionEncoding, SingleOwnerAccount};
-use starknet::core::types::FieldElement;
-use starknet::providers::jsonrpc::HttpTransport;
-use starknet::providers::JsonRpcClient;
-use starknet::signers::{LocalWallet, SigningKey};
-use tokio::sync::{Mutex, OnceCell};
-
-use crate::helpers::{chain_id, provider};
-use crate::katana_bench::N_TRANSACTIONS;
-use crate::ACCOUNT_ADDRESS;
-
-pub async fn account_manager() -> Arc<AccountManager> {
-    static CHAIN_ID: OnceCell<Arc<AccountManager>> = OnceCell::const_new();
-
-    CHAIN_ID
-        .get_or_init(|| async {
-            let mut accounts = AccountManager::generate().await;
-
-            let shared = accounts.remove(0); // remove the first account (it's the default account)
-            debug_assert_eq!(
-                shared.0.address(),
-                FieldElement::from_hex_be(ACCOUNT_ADDRESS).unwrap()
-            );
-
-            Arc::new(AccountManager { head: Arc::default(), accounts })
-        })
-        .await
-        .clone()
-}
-
-pub type OwnerAccount = SingleOwnerAccount<JsonRpcClient<HttpTransport>, LocalWallet>;
-pub type AccountWithNonce = (Arc<OwnerAccount>, Mutex<FieldElement>);
-
-pub struct AccountManager {
-    accounts: Vec<AccountWithNonce>,
-    head: Arc<Mutex<usize>>,
-}
-
-impl AccountManager {
-    async fn generate() -> Vec<AccountWithNonce> {
-        let mut seed = [0; 32];
-        seed[0] = 48;
-
-        let accounts =
-            DevAccountGenerator::new(N_TRANSACTIONS.try_into().unwrap()).with_seed(seed).generate();
-
-        let chain_id: FieldElement = chain_id().await;
-
-        let account_futures = accounts
-            .into_iter()
-            .map(|account| {
-                let private_key = SigningKey::from_secret_scalar(account.private_key);
-                let signer = LocalWallet::from_signing_key(private_key);
-                let account = SingleOwnerAccount::new(
-                    provider(),
-                    signer,
-                    account.address,
-                    chain_id,
-                    ExecutionEncoding::Legacy,
-                );
-
-                async move {
-                    let nonce = account.get_nonce().await.expect("Failed to get nonce");
-                    (Arc::new(account), Mutex::new(nonce))
-                }
-            })
-            .collect::<Vec<_>>();
-
-        join_all(account_futures).await
-    }
-
-    pub async fn next(&self) -> (Arc<OwnerAccount>, FieldElement) {
-        let mut head = self.head.lock().await;
-        *head = (*head + 1) % self.accounts.len();
-
-        let nonce_lock = self.accounts[*head].1.lock().await;
-        (self.accounts[*head].0.clone(), *nonce_lock + 1u8.into())
-    }
-}
